pyproject.toml
[project]
name = "investigate-fhe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "heir-py>=0.0.2",
]
